Help the Car go up/down hills
This simple update will allow the car to go at a controlled speed even when going up and down hills. The change included allowing the car to brake if its going too fast down a hill, another minor change was done in the simulator code to be able to receive braking info from the socket io connection. This change is important for the new track 2 which has lots of steep hills.